requirements.txt
Flask-Migrate
Flask-SQLAlchemy
Flask-WTF
flask-login